feat:IGED-182:  Factorie perform ui request used to display loading indicator after 300ms

+++ LaGrandeEcoleDuDroit/App/Presentation/Profile/AccountInformation/AccountInformationViewModel.swift
@@ -24,26 +24,7 @@ class AccountInformationViewModel: ViewModel {
     
     func updateProfilePicture(imageData: Data?) {
         guard let imageData else {
-            return event = ErrorEvent(message: "Image data is required.")
-        }
-        guard let user = uiState.user else { return }
-        
-        uiState.loading = true
-        
-        Task { @MainActor [weak self] in
-            do {
-                try await self?.updateProfilePictureUseCase.execute(user: user, imageData: imageData)
-                self?.resetScreenState()
-            } catch {
-                self?.resetScreenState()
-                self?.event = ErrorEvent(message: mapNetworkErrorMessage(error))
-            }
-        }
-    }
-    
-    func deleteProfilePicture() {
-        guard networkMonitor.isConnected else {
-            event = ErrorEvent(message: stringResource(.noInternetConectionError))
+            event = ErrorEvent(message: "Image data is required.")
             return
         }
         guard let user = uiState.user else {
@@ -51,21 +32,40 @@ class AccountInformationViewModel: ViewModel {
             return
         }
         
-        uiState.loading = true
+        performRequest { [weak self] in
+            try await self?.updateProfilePictureUseCase.execute(user: user, imageData: imageData)
+        }
+    }
+    
+    func deleteProfilePicture() {
+        guard let user = uiState.user else {
+            event = ErrorEvent(message: stringResource(.currentUserNotFoundError))
+            return
+        }
         
-        Task { @MainActor [weak self] in
-            do {
-                try await self?.userRepository.deleteProfilePicture(user: user)
-                self?.resetScreenState()
-            } catch {
-                self?.resetScreenState()
-                self?.event = ErrorEvent(message: mapNetworkErrorMessage(error))
-            }
+        performRequest { [weak self] in
+            try await self?.userRepository.deleteProfilePicture(user: user)
         }
     }
     
     func onScreenStateChange(_ state: ScreenState) {
         uiState.screenState = state
+    }
+    
+    private func performRequest(block: @escaping () async throws -> Void) {
+        performUiBlockingRequest(
+            block: block,
+            onLoading: { [weak self] in
+                self?.uiState.loading = true
+            },
+            onError: { [weak self] in
+                self?.event = ErrorEvent(message: mapNetworkErrorMessage($0))
+            },
+            onFinally: { [weak self] in
+                self?.uiState.loading = false
+                self?.resetScreenState()
+            }
+        )
     }
     
     private func initCurrentUser() {
